- Moving amqpproxy out of the root - Changing module name - it's not an Azure project (yet)
Things don't build yet, but this is just the pretty mechanical transform of the name.
