+++ go.mod
@@ -1,4 +1,4 @@
-module github.com/Azure/amqpfaultinjector
+module github.com/richardpark-msft/amqpfaultinjector
 
 go 1.23.0
