+++ internal/faultinjectors/slow_transfers_injector.go
@@ -5,9 +5,9 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/Azure/amqpfaultinjector/internal/logging"
-	"github.com/Azure/amqpfaultinjector/internal/proto/frames"
-	"github.com/Azure/amqpfaultinjector/internal/utils"
+	"github.com/richardpark-msft/amqpfaultinjector/internal/logging"
+	"github.com/richardpark-msft/amqpfaultinjector/internal/proto/frames"
+	"github.com/richardpark-msft/amqpfaultinjector/internal/utils"
 )
 
 // NewSlowTransfersInjector creates a SlowTransferFrames injector, which slows down any incoming
