+++ internal/faultinjectors/testbuffer_test.go
@@ -1,8 +1,8 @@
 package faultinjectors
 
 import (
-	"github.com/Azure/amqpfaultinjector/internal/proto/frames"
-	"github.com/Azure/amqpfaultinjector/internal/utils"
+	"github.com/richardpark-msft/amqpfaultinjector/internal/proto/frames"
+	"github.com/richardpark-msft/amqpfaultinjector/internal/utils"
 )
 
 type testBuffer struct {
