+++ internal/logging/json_logger_test.go
@@ -4,10 +4,10 @@ import (
 	"os"
 	"testing"
 
-	"github.com/Azure/amqpfaultinjector/internal/proto/frames"
-	"github.com/Azure/amqpfaultinjector/internal/proto/models"
-	"github.com/Azure/amqpfaultinjector/internal/utils"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
+	"github.com/richardpark-msft/amqpfaultinjector/internal/proto/frames"
+	"github.com/richardpark-msft/amqpfaultinjector/internal/proto/models"
+	"github.com/richardpark-msft/amqpfaultinjector/internal/utils"
 	"github.com/stretchr/testify/require"
 )
