+++ internal/proto/encoding/encode_test.go
@@ -3,7 +3,7 @@ package encoding
 import (
 	"testing"
 
-	"github.com/Azure/amqpfaultinjector/internal/proto/buffer"
+	"github.com/richardpark-msft/amqpfaultinjector/internal/proto/buffer"
 	"github.com/stretchr/testify/require"
 )
