+++ internal/proto/frames/bodies.go
@@ -6,8 +6,8 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/Azure/amqpfaultinjector/internal/proto/buffer"
-	"github.com/Azure/amqpfaultinjector/internal/proto/encoding"
+	"github.com/richardpark-msft/amqpfaultinjector/internal/proto/buffer"
+	"github.com/richardpark-msft/amqpfaultinjector/internal/proto/encoding"
 )
 
 // Type contains the values for a frame's type.
