+++ internal/proto/frames/frame.go
@@ -6,8 +6,8 @@ import (
 	"errors"
 	"math"
 
-	"github.com/Azure/amqpfaultinjector/internal/proto/buffer"
-	"github.com/Azure/amqpfaultinjector/internal/proto/encoding"
+	"github.com/richardpark-msft/amqpfaultinjector/internal/proto/buffer"
+	"github.com/richardpark-msft/amqpfaultinjector/internal/proto/encoding"
 )
 
 // PreambleOrFrame represents parseable items when extracting data from an AMQP
